Conflicts:
	plugins/versionpress/src/Storages/DirectoryStorage.php
	plugins/versionpress/src/Storages/PostMetaStorage.php
	plugins/versionpress/src/Utils/ArrayUtils.php

+++ plugins/versionpress/src/Database/MirroringDatabase.php
@@ -146,15 +146,23 @@ class MirroringDatabase extends ExtendedWpdb {
             }
 
             foreach ($ids as $id) {
+                $where['vp_id'] = $this->getVpId($entityName, $id);
+                if (!$where['vp_id']) {
+                    continue; // already deleted - deleting postmeta is sometimes called twice
+                }
+
                 if ($entityName === 'postmeta' && !isset($where['vp_post_id'])) {
                     $where['vp_post_id'] = $this->get_var("select HEX(reference_vp_id) from {$this->dbSchemaInfo->getPrefixedTableName('vp_reference_details')} where `table` = 'postmeta' and id = " . $where[$idColumnName]);
+                }
+
+                if ($entityName === 'usermeta' && !isset($where['vp_user_id'])) {
+                    $where['vp_user_id'] = $this->get_var("select HEX(reference_vp_id) from {$this->dbSchemaInfo->getPrefixedTableName('vp_reference_details')} where `table` = 'usermeta' and id = " . $where[$idColumnName]);
                 }
 
                 if ($hasReferences) {
                     $this->deleteReferences($entityName, $id);
                 }
 
-                $where['vp_id'] = $this->getVpId($entityName, $id);
                 $this->deleteId($entityName, $id);
                 $this->mirror->delete($entityName, $where);
             }
